Override GetDefaultUserDataDirectory() in chrome_paths_linux.cc
