+++ common/BUILD.gn
@@ -50,9 +50,3 @@ source_set("common") {
   }
 }
 
-source_set("constants") {
-  sources = [
-    "//brave/common/brave_channel_info_linux.cc",
-    "//brave/common/brave_channel_info_linux.h",
-  ]
-}
