--- common/brave_channel_info_linux.h
@@ -1,16 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#ifndef BRAVE_COMMON_BRAVE_CHANNEL_INFO_LINUX_H_
-#define BRAVE_COMMON_BRAVE_CHANNEL_INFO_LINUX_H_
-
-#include <string>
-
-namespace brave {
-
-std::string GetChannelSuffixForDataDir();
-
-}  // namespace brave
-
-#endif  // BRAVE_COMMON_BRAVE_CHANNEL_INFO_LINUX_H_
