--- patches/chrome-common-BUILD.gn.patch
@@ -1,17 +1,0 @@
-diff --git a/chrome/common/BUILD.gn b/chrome/common/BUILD.gn
-index eb650f8aa7009371e31077e0f4a754807efd7613..11c036a1922da9f00a8b196657bdfc15431c4f9c 100644
---- a/chrome/common/BUILD.gn
-+++ b/chrome/common/BUILD.gn
-@@ -579,6 +579,12 @@ static_library("constants") {
-     "//third_party/widevine/cdm:headers",
-   ]
- 
-+  if (brave_chromium_build) {
-+    deps += [
-+      "//brave/common:constants"
-+    ]
-+  }
-+
-   if (is_android) {
-     # This dependency must only be added for Android.
-     #
