--- patches/chrome-common-chrome_paths_linux.cc.patch
@@ -1,30 +1,0 @@
-diff --git a/chrome/common/chrome_paths_linux.cc b/chrome/common/chrome_paths_linux.cc
-index fc47bd3f12418fdaed879e0b485bfc7cb572a6e8..30585ca8b71bb0a7c0e3c817ea15493f4a4212bc 100644
---- a/chrome/common/chrome_paths_linux.cc
-+++ b/chrome/common/chrome_paths_linux.cc
-@@ -16,6 +16,10 @@
- #include "chrome/common/channel_info.h"
- #include "chrome/common/chrome_paths_internal.h"
- 
-+#if defined(BRAVE_CHROMIUM_BUILD)
-+#include "brave/common/brave_channel_info_linux.h"
-+#endif
-+
- namespace chrome {
- 
- using base::nix::GetXDGDirectory;
-@@ -87,10 +91,14 @@ bool GetDefaultUserDataDirectory(base::FilePath* result) {
-         GetXDGDirectory(env.get(), kXdgConfigHomeEnvVar, kDotConfigDir);
-   }
- 
-+#if defined(BRAVE_CHROMIUM_BUILD)
-+  *result = config_dir.Append("BraveSoftware/Brave-Browser" + brave::GetChannelSuffixForDataDir());
-+#else
- #if defined(GOOGLE_CHROME_BUILD)
-   *result = config_dir.Append("google-chrome" + GetChannelSuffixForDataDir());
- #else
-   *result = config_dir.Append("chromium");
-+#endif
- #endif
-   return true;
- }
